Modifications to bring in the new serialiser: - Switched to 64 bit lengths for files. - Switched to new data types (RsItem). - Moved Msg and Chat to services.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@275 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/server/ft.cc
@@ -26,7 +26,7 @@
 
 #include "server/ft.h"
 
-bool	ftManager::lookupLocalHash(std::string hash, std::string &path, uint32_t &size)
+bool	ftManager::lookupLocalHash(std::string hash, std::string &path, uint64_t &size)
 {
 	std::list<FileDetail> details;
